% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{at_home}
\alias{at_home}
\title{Detect not on CRANity}
\usage{
at_home()
}
\description{
Detect whether we are running at home (i.e. not on CRAN, BioConductor, ...)
}
\examples{
# test will run locally, but not on CRAN
if ( at_home() ){
  expect_equal(2, 1+1)
}
}
\concept{test-functions test-file}
