% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expect_length}
\alias{expect_length}
\title{Check length of an object}
\usage{
expect_length(current, length, info = NA_character_, ...)
}
\arguments{
\item{current}{\code{[object]} An R object with a length}

\item{length}{\code{[integer]} A nonnegative integer}

\item{info}{\code{[character]} scalar. Optional user-defined message. Must
be a single character string. Multiline comments may be separated by}

\item{...}{Currently not used.}
}
\description{
Check length of an object
}
\examples{
expect_length(3:4, 2) # TRUE
expect_length(2:5, 1) # FALSE

}
\seealso{
Other test-functions: 
\code{\link{expect_equal}()},
\code{\link{expect_equal_to_reference}()},
\code{\link{expect_match}()},
\code{\link{ignore}()}
}
\concept{test-functions}
