% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{get_call_wd}
\alias{get_call_wd}
\title{Get workding dir from where a test was initiated}
\usage{
get_call_wd()
}
\value{
\code{[character]} A path.
}
\description{
A test runner, like \code{\link{run_test_file}} changes
R's working directory to the location of the test file temporarily 
while the tests run. This function can be used from within the
test file to get R's working directory at the time \code{run_test_file} 
(or one of it's siblings)
was called.
}
\examples{
get_call_wd()
}
