% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{format.tinytest}
\alias{format.tinytest}
\alias{print.tinytest}
\title{Print a tinytest object}
\usage{
\method{format}{tinytest}(x, type = c("long", "short"), ...)

\method{print}{tinytest}(x, ...)
}
\arguments{
\item{x}{A \code{tinytest} object}

\item{type}{\code{[logical]} Toggle format type}

\item{...}{passed to \code{\link{format.tinytest}}}
}
\value{
A character string
}
\description{
Print a tinytest object
}
\examples{
tt <- expect_equal(1+1, 3)
format(tt,"long")
format(tt,"short")
print(expect_equal(1+1, 2))
print(expect_equal(1+1, 3), type="long")

}
