% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{puppy}
\alias{puppy}
\title{The puppy for a pkgKitten}
\usage{
puppy(pkgdir, force = FALSE, verbose = TRUE)
}
\arguments{
\item{pkgdir}{\code{[character]} Package source directory}

\item{force}{\code{[logical]} Toggle overwrite existing files? (not folders)}

\item{verbose}{\code{[logical]} Toggle print progress}
}
\description{
Does exactly the same as \code{\link{setup_tinytest}}, but prints 
a loving message aferwards (and who doesn't want that!?). Just
think about those puppies.
}
\keyword{internal}
