% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_tinytest}
\alias{setup_tinytest}
\title{Add tinytest to package source directory}
\usage{
setup_tinytest(pkgdir, force = FALSE, verbose = TRUE)
}
\arguments{
\item{pkgdir}{\code{[character]} Package source directory}

\item{force}{\code{[logical]} Toggle overwrite existing files? (not folders)}

\item{verbose}{\code{[logical]} Toggle print progress}
}
\value{
\code{NULL}, invisibly.
}
\description{
Creates \code{inst/tinytest}, and an example test file in that
directory. Creates \code{tests/tinytest.R} so the package is
tested with \code{R CMD check}. Adds \code{tinytests} as a suggested
package to the \code{DESCRIPTION}.
}
\section{Note on \code{DESCRIPTION}}{


Fails when it does not exist. It is assumed that the
package is named in the \code{DESCRIPTION}.
}

\examples{
\dontrun{
# an easy way to set up a package 'haha' that passes
# R CMD check
pkgKitten::kitten("haha")
tinytest::setup_tinytest("haha")
}

}
