% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{exit_file}
\alias{exit_file}
\title{Stop testing}
\usage{
exit_file(msg = "")
}
\arguments{
\item{msg}{\code{[character]} An optional message to print after exiting.}
}
\value{
The exit message
}
\description{
Call this function to exit a test file.
}
\examples{
exit_file("I'm too tired to test")

}
\seealso{
Other test-files: 
\code{\link{build_install_test}()},
\code{\link{run_test_dir}()},
\code{\link{run_test_file}()},
\code{\link{summary.tinytests}()},
\code{\link{test_package}()}
}
\concept{test-files}
